Refactor WorkflowStateSyncer to use the ResponseSyncerSender and ResponseSyncerReceiver classes
