+++ opengever/task/locales/de/LC_MESSAGES/opengever.task.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: 1.0\n"
-"POT-Creation-Date: 2017-03-23 08:33+0000\n"
+"POT-Creation-Date: 2017-04-12 09:31+0000\n"
 "PO-Revision-Date: 2015-02-17 18:41+0100\n"
 "Last-Translator: Jonas Baumann <j.baumann@4teamwork.ch>\n"
 "Language-Team: 4teamwork <info@4teamwork.ch>\n"
@@ -22,7 +22,7 @@ msgstr "${num} Dokumente wurden kopiert."
 msgid "${subtask_num} subtasks were create."
 msgstr "${subtask_num} Unteraufgaben wurden erstellt."
 
-#: ./opengever/task/response.py:434
+#: ./opengever/task/response.py:402
 msgid "Changes saved to response id ${response_id}."
 msgstr "Änderungen an der Antwort ${response_id} wurden gespeichert"
 
@@ -34,29 +34,29 @@ msgstr "Löschen"
 msgid "Edit"
 msgstr "Bearbeiten"
 
-#: ./opengever/task/response.py:337
-#: ./opengever/task/statesyncer.py:145
+#: ./opengever/task/response.py:305
+#: ./opengever/task/syncer.py:219
 #: ./opengever/task/util.py:162
 msgid "Issue state"
 msgstr "Status"
 
-#: ./opengever/task/response.py:462
+#: ./opengever/task/response.py:430
 msgid "No response selected for removal."
 msgstr "Es wurde keine Antwort fürs Löschen selektiert"
 
-#: ./opengever/task/response.py:425
+#: ./opengever/task/response.py:393
 msgid "No response selected for saving."
 msgstr "Es wurde keine Antwort fürs Speichern selektiert"
 
-#: ./opengever/task/response.py:484
+#: ./opengever/task/response.py:452
 msgid "Removed response id ${response_id}."
 msgstr "Antwort ${response_id} gelöscht"
 
-#: ./opengever/task/response.py:477
+#: ./opengever/task/response.py:445
 msgid "Response id ${response_id} does not exist so it cannot be removed."
 msgstr "Antowrt ${response_id} existiert nicht und konnte deshalb nicht gelöscht werden"
 
-#: ./opengever/task/response.py:469
+#: ./opengever/task/response.py:437
 msgid "Response id ${response_id} is no integer so it cannot be removed."
 msgstr "Antwort id ${response_id} ist keine Zahl und konnte deshalb nicht gelöscht werden"
 
@@ -105,11 +105,11 @@ msgstr "Die Aufgabe wurde akzeptiert."
 msgid "The task has been copied to the selected dossier and accepted."
 msgstr "Die Aufgabe wurde kopiert und akzeptiert."
 
-#: ./opengever/task/response.py:456
+#: ./opengever/task/response.py:424
 msgid "You are not allowed to delete responses."
 msgstr "Sie haben keine Berechtigung Antworten zu löschen"
 
-#: ./opengever/task/response.py:419
+#: ./opengever/task/response.py:387
 msgid "You are not allowed to edit responses."
 msgstr "Sie haben keine Berechtigung Antworten zu editieren"
 
@@ -187,11 +187,11 @@ msgid "button_save"
 msgstr "Speichern"
 
 #. Default: "Cancel"
-#: ./opengever/task/response.py:181
+#: ./opengever/task/response.py:179
 msgid "cancel"
 msgstr "Abbrechen"
 
-#: ./opengever/task/response.py:297
+#: ./opengever/task/response.py:265
 msgid "date_of_completion"
 msgstr "Erledigungsdatum"
 
@@ -286,7 +286,7 @@ msgstr "Wählen Sie die Dokumente aus, die Sie dem Auftraggeber liefern wollen. 
 msgid "help_complete_task_response"
 msgstr "Der hier eingegebene Text erscheint in der Antwort, die die Aufgabe abschliesst."
 
-#: ./opengever/task/response.py:77
+#: ./opengever/task/response.py:69
 #: ./opengever/task/task.py:139
 msgid "help_date_of_completion"
 msgstr "Das Datum an dem die Aufgabe beendet wurde"
@@ -415,13 +415,13 @@ msgstr "erstellt am"
 
 #. Default: "Date of completion"
 #: ./opengever/task/browser/overview.py:103
-#: ./opengever/task/response.py:76
+#: ./opengever/task/response.py:68
 #: ./opengever/task/task.py:138
 msgid "label_date_of_completion"
 msgstr "Erledigungsdatum"
 
 #. Default: "Deadline"
-#: ./opengever/task/activities.py:93
+#: ./opengever/task/activities.py:57
 #: ./opengever/task/browser/delegate/metadata.py:32
 #: ./opengever/task/browser/overview.py:87
 msgid "label_deadline"
@@ -433,7 +433,7 @@ msgid "label_documents"
 msgstr "Dokumente"
 
 #. Default: "Dossier title"
-#: ./opengever/task/activities.py:97
+#: ./opengever/task/activities.py:61
 msgid "label_dossier_title"
 msgstr "Dossiertitel"
 
@@ -511,7 +511,7 @@ msgstr "Ursprüngliche Aufgabe"
 
 #. Default: "Related Items"
 #: ./opengever/task/browser/complete.py:214
-#: ./opengever/task/response.py:53
+#: ./opengever/task/response.py:74
 #: ./opengever/task/task.py:152
 msgid "label_related_items"
 msgstr "Verweise"
@@ -532,7 +532,7 @@ msgstr "Antworttext"
 #. Default: "Responsible"
 #: ./opengever/task/browser/assign.py:44
 #: ./opengever/task/browser/overview.py:98
-#: ./opengever/task/statesyncer.py:136
+#: ./opengever/task/syncer.py:210
 msgid "label_responsible"
 msgstr "Auftragnehmer"
 
@@ -557,7 +557,7 @@ msgid "label_successor_task"
 msgstr "Kopierte Aufgabe"
 
 #. Default: "New task added by ${user}"
-#: ./opengever/task/activities.py:67
+#: ./opengever/task/activities.py:31
 msgid "label_task_added"
 msgstr "Neue Aufgabe hinzugefügt durch ${user}"
 
@@ -567,20 +567,20 @@ msgid "label_task_commented"
 msgstr "Aufgabe wurde kommentiert"
 
 #. Default: "Task title"
-#: ./opengever/task/activities.py:92
+#: ./opengever/task/activities.py:56
 #: ./opengever/task/browser/overview.py:66
 msgid "label_task_title"
 msgstr "Aufgabentitel"
 
 #. Default: "Task Type"
-#: ./opengever/task/activities.py:95
+#: ./opengever/task/activities.py:59
 #: ./opengever/task/browser/overview.py:78
 #: ./opengever/task/task.py:102
 msgid "label_task_type"
 msgstr "Auftragstyp"
 
 #. Default: "Text"
-#: ./opengever/task/activities.py:99
+#: ./opengever/task/activities.py:63
 #: ./opengever/task/browser/delegate/metadata.py:37
 #: ./opengever/task/browser/overview.py:74
 msgid "label_text"
@@ -643,7 +643,7 @@ msgid "same_deadline_error"
 msgstr "Die angegebene Frist ist die aktuelle."
 
 #. Default: "Save"
-#: ./opengever/task/response.py:162
+#: ./opengever/task/response.py:166
 msgid "save"
 msgstr "Speichern"
 
@@ -767,7 +767,7 @@ msgid "transition_label_created"
 msgstr "Erstellt durch ${user}"
 
 #. Default: "Task added"
-#: ./opengever/task/activities.py:62
+#: ./opengever/task/activities.py:26
 #: ./opengever/task/response_description.py:49
 msgid "transition_label_default"
 msgstr "Aufgabe hinzugefügt"
